refactor PrintNode request
