+++ app/Modules/PrintNode/src/Http/Requests/PrintJobStoreRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Modules\PrintNode\src\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StorePrintJobRequest extends FormRequest
+class PrintJobStoreRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
