+++ app/Modules/PrintNode/src/Http/Requests/PrintNodeIndexRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\PrintNode\PrinterController;
+namespace App\Modules\PrintNode\src\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class IndexRequest extends FormRequest
+class PrintNodeIndexRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
